chore: move infra test utils from test to common
commit-id:68fd1fa2
